Added TerminationResistance property to CanNetworkInterface class - closes #98

+++ src/SocketCANSharp/Network/Netlink/CanDeviceProperties.cs
@@ -73,6 +73,6 @@ namespace SocketCANSharp.Network.Netlink
         /// <summary>
         /// Switchable Termination Resistance option.
         /// </summary>
-        public uint? TerminationResistance { get; set; }
+        public ushort? TerminationResistance { get; set; }
     }
 }
